+++ CHANGELOG.md
@@ -6,6 +6,7 @@ All notable changes will be documented in this file.
 
 ### Changed
 
+* `string_buffer` did not have O(1) `length()` (#53)
 * Move OMS interfaces to API (#51)
 * `RateLimitTrigger` must support const members (#34)